Make opensubdiv-petite work in windows

+++ opensubdiv-petite-sys/src/far/topology_refiner.rs
@@ -5,11 +5,12 @@ use num_enum::TryFromPrimitive;
 #[derive(TryFromPrimitive, Copy, Clone, Debug)]
 pub enum Scheme {
     /// *Bilinear* interpolation.
-    Bilinear = crate::OpenSubdiv_v3_4_4_Sdc_SchemeType_SCHEME_BILINEAR,
+    Bilinear = crate::OpenSubdiv_v3_4_4_Sdc_SchemeType_SCHEME_BILINEAR as u32,
     /// [*Catmull-Clark* subdivision](https://en.wikipedia.org/wiki/Catmull%E2%80%93Clark_subdivision_surface).
-    CatmullClark = crate::OpenSubdiv_v3_4_4_Sdc_SchemeType_SCHEME_CATMARK,
+    CatmullClark =
+        crate::OpenSubdiv_v3_4_4_Sdc_SchemeType_SCHEME_CATMARK as u32,
     /// [*Loop* subdivision](https://en.wikipedia.org/wiki/Loop_subdivision_surface).
-    Loop = crate::OpenSubdiv_v3_4_4_Sdc_SchemeType_SCHEME_LOOP,
+    Loop = crate::OpenSubdiv_v3_4_4_Sdc_SchemeType_SCHEME_LOOP as u32,
 }
 
 #[repr(u32)]
@@ -22,31 +23,31 @@ pub enum BoundaryInterpolation {
     ///
     /// Boundary faces can be selectively interpolated by sharpening all
     /// boundary edges incident the vertices of the face.
-    None = crate::OpenSubdiv_v3_4_4_Sdc_Options_VtxBoundaryInterpolation_VTX_BOUNDARY_NONE,
+    None = crate::OpenSubdiv_v3_4_4_Sdc_Options_VtxBoundaryInterpolation_VTX_BOUNDARY_NONE as u32,
     /// A sequence of boundary vertices defines a smooth curve to which the
     /// limit surface along boundary faces extends.
-    EdgeOnly = crate::OpenSubdiv_v3_4_4_Sdc_Options_VtxBoundaryInterpolation_VTX_BOUNDARY_EDGE_ONLY,
+    EdgeOnly = crate::OpenSubdiv_v3_4_4_Sdc_Options_VtxBoundaryInterpolation_VTX_BOUNDARY_EDGE_ONLY as u32,
     /// Similar to edge-only but the smooth curve resulting on the boundary is
     /// made to interpolate corner vertices (vertices with exactly one incident
     /// face).
-    EdgeAndCorner = crate::OpenSubdiv_v3_4_4_Sdc_Options_VtxBoundaryInterpolation_VTX_BOUNDARY_EDGE_AND_CORNER,
+    EdgeAndCorner = crate::OpenSubdiv_v3_4_4_Sdc_Options_VtxBoundaryInterpolation_VTX_BOUNDARY_EDGE_AND_CORNER as u32,
 }
 
 #[repr(u32)]
 #[derive(TryFromPrimitive, Copy, Clone, Debug)]
 pub enum FaceVaryingLinearInterpolation {
     /// Smooth everywhere the mesh is smooth.
-    None = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_NONE,
+    None = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_NONE as u32,
     /// Linearly interpolate (sharpen or pin) corners only,
-    CornersOnly = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_CORNERS_ONLY,
+    CornersOnly = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_CORNERS_ONLY as u32,
     /// `CornersOnly` + sharpening of junctions of three or more regions.
-    CornersPlusOne = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_CORNERS_PLUS1,
+    CornersPlusOne = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_CORNERS_PLUS1 as u32,
     /// `CornersPlusOne` + sharpening of darts and concave corners.
-    CornersPlusTwo = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_CORNERS_PLUS2,
+    CornersPlusTwo = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_CORNERS_PLUS2 as u32,
     /// `Linear interpolation along all boundary edges and corners.
-    Boundaries = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_BOUNDARIES,
+    Boundaries = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_BOUNDARIES as u32,
     /// Linear interpolation everywhere (boundaries and interior).
-    All = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_ALL,
+    All = crate::OpenSubdiv_v3_4_4_Sdc_Options_FVarLinearInterpolation_FVAR_LINEAR_ALL as u32,
 }
 
 #[repr(u32)]
@@ -55,8 +56,8 @@ pub enum CreasingMethod {
     /// Apply regular, *Catmull-Clark* semi-sharp crease rules.
     ///
     /// * Note that this may give a jagged look when crease values vary along an [edge loop](https://en.wikipedia.org/wiki/Edge_loop).
-    Uniform =
-        crate::OpenSubdiv_v3_4_4_Sdc_Options_CreasingMethod_CREASE_UNIFORM,
+    Uniform = crate::OpenSubdiv_v3_4_4_Sdc_Options_CreasingMethod_CREASE_UNIFORM
+        as u32,
     /// Apply *Chaikin* semi-sharp crease rules.
     ///
     /// The *Chaikin Rule* is a variation of the semi-sharp creasing method
@@ -65,19 +66,22 @@ pub enum CreasingMethod {
     /// the subdivision of sharpness values using Chaikin's curve subdivision
     /// algorithm to consider all sharpness values of edges around a common
     /// vertex when determining the sharpness of child edges.
-    Chaikin =
-        crate::OpenSubdiv_v3_4_4_Sdc_Options_CreasingMethod_CREASE_CHAIKIN,
+    Chaikin = crate::OpenSubdiv_v3_4_4_Sdc_Options_CreasingMethod_CREASE_CHAIKIN
+        as u32,
 }
 
 #[repr(u32)]
 #[derive(TryFromPrimitive, Copy, Clone, Debug)]
 pub enum TriangleSubdivision {
     /// Default *Catmull-Clark* scheme weights at triangles.
-    CatmullClark = crate::OpenSubdiv_v3_4_4_Sdc_Options_TriangleSubdivision_TRI_SUB_CATMARK,
+    CatmullClark =
+        crate::OpenSubdiv_v3_4_4_Sdc_Options_TriangleSubdivision_TRI_SUB_CATMARK
+            as u32,
     /// Modifies the subdivision behavior at triangular faces to improve the
     /// undesirable surface artefacts that often result in such areas.
     Smooth =
-        crate::OpenSubdiv_v3_4_4_Sdc_Options_TriangleSubdivision_TRI_SUB_SMOOTH,
+        crate::OpenSubdiv_v3_4_4_Sdc_Options_TriangleSubdivision_TRI_SUB_SMOOTH
+            as u32,
 }
 
 pub type UniformRefinementOptions =
